Create base-test jobs for staging changes
Because it is possible to break zuul with changes to base, we really
need a way to properly stage these jobs without breaking running jobs.

This is a process we currently use in openstack-infra and has been
very helpful to catching issues.

We also add a noop job to gate, to allow zuul to merge changes to
zuul-config.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>