Secrets: Provide encrypted keys w/o \n
The keys that were provided were encrypted with a \n in the file (vim
default) This lead to the access key being "access_key\n" which is not a
valid access_key.

Newly generated content have been generated with a vim file configured
to not have a \n (:set noeol, :set binary) the content of the access key
should be correct now.