gate: remove clean-check requirement
Clean-check means that gate does not start without a successful check.
Turns out that is a requirement that has been lifted from the openstack old days, and it we could remove it.
Accepting this change results in:

once you approve and add the gate-it tag, then check jobs are canceled and the change goes directly to gate.